change activeContractID to Long in CustomerFullDetailsRecord

+++ Backend/src/main/server/sql/dto/customer/CustomerFullDetailsRecord.java
@@ -4,7 +4,7 @@ import main.server.sql.entities.eCustomerStatus;
 
 public record CustomerFullDetailsRecord(
 		Short customerID,
-		Integer activeContractID,
+		Long activeContractID,
 		String customerShortName,
 		String customerName,
 		eCustomerStatus customerStatus,
